Changed some of my code to maintain consistency, also fixed an incorrect data type in BankAccount. Improved the Name class' validation methods.

+++ src/code/ca/bcit/comp2522/bank/BankAccount.java
@@ -30,7 +30,7 @@ public class BankAccount
      * @param accountOpenedDate a Date this person opened account
      * @param accountClosedDate a Date this person closed account
      */
-    public BankAccount(final BankClient client, final String accountNumber, final int pin, final int balanceUsd, final Date accountOpenedDate, final Date accountClosedDate)
+    public BankAccount(final BankClient client, final String accountNumber, final int pin, final double balanceUsd, final Date accountOpenedDate, final Date accountClosedDate)
     {
         // assign variables
         this.client = client;
